refactor: move UI files to better namespaces
