+++ CommunityProject/Assets/_Game/Scripts/Runtime/UI/Credits/CreditsTextBuilder.cs.meta
